Integrate Schema with Backend
Integrate working schema with the backend by changing the mapping in the
AppUser, AppRole and UserRole entities to match the current schema.

Login functionality is currently working.

+++ src/main/java/edu/ccsu/datamail/jobtracker/dao/AppRoleDAO.java
@@ -17,7 +17,7 @@ public class AppRoleDAO
     @Autowired
     private EntityManager entityManager;
 
-    public List<String> getRoleNames(Long userId)
+    public List<String> getRoleNames(Integer userId)
     {
         String sql = "Select ur.appRole.roleName from " + UserRole.class.getName() + " ur " //
                 + " where ur.appUser.userId = :userId ";
